Box blur now works

+++ src/fxblur.h
@@ -44,7 +44,7 @@ public:
 private:
   void boxBlur(QRgb *buffer1, QRgb *buffer2, int width, int height, int radius);
   void boxBlurHorizontal(QRgb *buffer1, QRgb *buffer2, int width, int height, int radius);
-  void boxBlurTotal(QRgb *buffer1, QRgb *buffer2, int width, int height, double radius);
+  void boxBlurTotal(QRgb *buffer1, QRgb *buffer2, int width, int height, int radius);
   
 };
